fix : PK 자료형 바꾼 후 오류 수정

+++ src/main/java/com/ll/gramgram/boundedContext/likeablePerson/service/LikeablePersonService.java
@@ -50,7 +50,7 @@ public class LikeablePersonService {
     }
 
     @Transactional
-    public RsData<LikeablePerson> delete(int id) {
+    public RsData<LikeablePerson> delete(Long id) {
         likeablePersonRepository.deleteById(id);
         return RsData.of("S-1", "선택하신 인스타유저가 호감상대에서 삭제되었습니다.");
     }
